refactor/improve home and about
