+++ components/carousel/index.tsx
@@ -1,4 +1,5 @@
 import Image from 'next/image';
+import React from 'react';
 import Carousel from 'react-multi-carousel';
 import 'react-multi-carousel/lib/styles.css';
 
@@ -64,4 +65,4 @@ const NiftyCarousel = ({
   </>
 );
 
-export default NiftyCarousel;
+export default React.memo(NiftyCarousel);
